+++ src/pages/UserProfile/Settings/EmailSetter/EmailSetter.module.scss
@@ -1,0 +1,13 @@
+@import "../../../../assets/scss/widths";
+
+.emailChangeForm {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: 1fr 1fr;
+  column-gap: 1rem;
+  row-gap: 2rem;
+
+  @media (min-width: $break-md) {
+    grid-template-columns: 1fr 1fr;
+  }
+}
