+++ src/pages/UserProfile/Settings/PasswordSetter/PasswordSetter.module.scss
@@ -1,0 +1,18 @@
+@import "../../../../assets/scss/widths";
+
+.passwordChangeForm {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-template-rows: 1fr 1fr;
+  column-gap: 1rem;
+  row-gap: 2rem;
+  & .invisible {
+    display: none;
+  }
+  @media (min-width: $break-md) {
+    grid-template-columns: 1fr 1fr;
+    & .invisible {
+      display: block;
+    }
+  }
+}
