+++ src/pages/UserProfile/Settings/Settings.module.scss
@@ -36,42 +36,3 @@
     }
   }
 }
-
-.passwordChangeForm {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr 1fr;
-  column-gap: 1rem;
-  row-gap: 2rem;
-  & .invisible {
-    display: none;
-  }
-  @media (min-width: $break-md) {
-    grid-template-columns: 1fr 1fr;
-    & .invisible {
-      display: block;
-    }
-  }
-}
-
-.bankForm {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr 1fr;
-  column-gap: 1rem;
-  row-gap: 2rem;
-  align-items: center;
-  & .currentlyOnly {
-    height: 100%;
-    background-color: #beeeee;
-    color: #007878;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 14px;
-    line-height: 16px;
-  }
-  @media (min-width: $break-md) {
-    grid-template-columns: 1fr 1fr;
-  }
-}
